requirements.txt
numpy
pandas
dash_html_components
dash_core_components
dash_table
plotly
sqlalchemy
